Move string decoding utilities to `common`
